feat: 상품 옵션 기능 추가
- ProductOption 엔티티 추가 및 Product와 연관관계 매핑
- 장바구니/주문/결제 흐름에 선택 옵션 및 추가금액 반영
- 결제 완료 시 옵션별/상품별 재고 차감 처리

+++ src/main/java/com/commerce/admin/dto/AdminProductListDTO.java
@@ -16,7 +16,7 @@ public class AdminProductListDTO {
 	private Long id;
 	private String name;
 	private int price;
-	private int stock;
+	private long stock;
 	private String mainImageUrl;
 	private LocalDateTime createdAt;
